Rewrite test cases
