+++ tests/District-test.js
@@ -1,28 +1,29 @@
-import '../.auto_mock_off';
-import React, { findDOMNode } from 'react/addons';
-import TWzipcode from '../src/containers/App';
+import React from 'react';
+import { findDOMNode } from 'react-dom';
+import TestUtils from 'react-addons-test-utils';
+import { expect } from 'chai';
+import TWzipcode from '../src';
 import District from '../src/components/District';
 
 describe('County', () => {
-  var TestUtils = React.addons.TestUtils;
-  var component;
+  let component;
 
   beforeEach(() => {
     component = TestUtils.renderIntoDocument(
       <TWzipcode css={['form-control county-sel',
                        'form-control district-sel',
                        'form-control zipcode']}
-                 zipcodeSel={'251'} />
+                 zipcodeValue={'251'} />
     );
   });
 
   it('set zipcode as 251, then district should be "淡水區"', () => {
     let district = TestUtils.findRenderedComponentWithType(component, District);
-    expect(district.props.defaultValue).toEqual('淡水區');
+    expect(district.props.value).to.equal('淡水區');
   });
 
   it('set district className', () => {
     let district = TestUtils.findRenderedComponentWithType(component, District);
-    expect(findDOMNode(district).className).toEqual('form-control district-sel');
+    expect(findDOMNode(district).className).to.equal('form-control district-sel');
   });
 });
