+++ tests/setup.js
@@ -1,0 +1,18 @@
+import jsdom from 'jsdom';
+import sinon from 'sinon';
+
+// Define some html to be our basic document
+// JSDOM will consume this and act as if we were in a browser
+const DEFAULT_HTML = '<html><body></body></html>';
+
+// Define some variables to make it look like we're a browser
+// First, use JSDOM's fake DOM as the document
+global.document = jsdom.jsdom(DEFAULT_HTML);
+
+// Set up a mock window
+global.window = document.defaultView;
+
+// Allow for things like window.location
+global.navigator = window.navigator;
+
+global.XMLHttpRequest = sinon.useFakeXMLHttpRequest();
